pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "jpg2xlsx"
version = "0.1.0"
description = "Convert images into Excel spreadsheets where each cell matches a pixel color."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Tlaloc-Es" }]
keywords = ["cli", "excel", "image", "jpg", "xlsx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Office/Business :: Financial :: Spreadsheet",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = ["Pillow>=10.0.0", "XlsxWriter>=3.2.0"]

[project.optional-dependencies]

[project.urls]
Homepage = "https://github.com/Tlaloc-Es/jpg2xlsx"
Repository = "https://github.com/Tlaloc-Es/jpg2xlsx"
Issues = "https://github.com/Tlaloc-Es/jpg2xlsx/issues"
Changelog = "https://github.com/Tlaloc-Es/jpg2xlsx/blob/master/CHANGELOG.md"

[project.scripts]
jpg2xlsx = "jpg2xlsx.cli:main"

[dependency-groups]
dev = [
    "commitizen>=3.29.0",
    "openpyxl>=3.1.5",
    "pre-commit>=4.2.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/jpg2xlsx"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
version_files = [
    "pyproject.toml:version",
    "src/jpg2xlsx/__init__.py:__version__",
]
